fix(settings) GBP-29: Fix some folder settings
