+++ src/app/shared/general.model.ts
