+++ src/index.html
@@ -15,9 +15,7 @@
 </head>
 
 <body class="container">
-  <header>Header</header>
   <app-root></app-root>
-  <footer>Footer</footer>
 </body>
 
 </html>
