+++ src/stories/dropdown/dropdown.component.ts
@@ -1,5 +1,5 @@
 import { Component, EventEmitter, Input, Output } from '@angular/core';
-import { DropdownValue } from '../../shared/general.model';
+import { DropdownValue } from '../../app/shared/general.model';
 
 @Component({
   selector: 'sb-dropdown',
